Working on removing warnings related to sizes on 64 bit architectures.
  1. In afun.h and afun.c
      - Changed return type of AT_symbolTableSize to size_t (from unsigned int)
      - Changed cast of Symbol to size_t (from unsigned int) so that
        comparision with AT_symbolTableSize doesn't overflow.  This fixes a
        major bug and allows the fib.c test to complete successfully!
  2. Made failure to set SIZEOF_INT to 4 or 8 a compiler error

+++ aterm/aterm/afun.c
@@ -98,9 +98,9 @@ static void resize_table()
 
 /*}}}  */
 
-/*{{{  unsigned int AT_symbolTableSize() */
+/*{{{  size_t AT_symbolTableSize() */
 
-unsigned int AT_symbolTableSize()
+size_t AT_symbolTableSize()
 {
   return table_size;
 }
